delivery_backend/requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
environs==9.3.2
flake8==3.9.1
gunicorn==20.1.0
marshmallow==3.11.1
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.8.6
dj-database-url==0.5.0
geopy==2.1.0
django-rest-framework-braces==0.3.4
